+++ apps/docs/src/app/core/api-files.ts
@@ -289,5 +289,6 @@ export const API_FILES = {
         'ToolbarOverflowContentComponent',
         'ToolbarSpacerComponent'
     ],
-    tree: ['TreeComponent', 'TreeChildComponent', 'TreeRowObjectModel']
+    tree: ['TreeComponent', 'TreeChildComponent', 'TreeRowObjectModel'],
+    wizard: ['WizardComponent']
 };