+++ apps/docs/src/app/core/component-docs/wizard/examples/wizard-example.component.ts
@@ -1,0 +1,50 @@
+import { Component, ViewEncapsulation } from '@angular/core';
+import { WizardStepStatus } from '@fundamental-ngx/core';
+
+@Component({
+    selector: 'fd-wizard-example',
+    templateUrl: './wizard-example.component.html',
+    encapsulation: ViewEncapsulation.None,
+    styles: [
+        `
+            .fd-wizard-example .fd-wizard__content {
+                min-height: 300px;
+            }
+        `
+    ],
+    host: {
+        class: 'fd-wizard-example'
+    }
+})
+export class WizardExampleComponent {
+    step1status: WizardStepStatus = 'current';
+    step2status: WizardStepStatus = 'upcoming';
+    step3status: WizardStepStatus = 'upcoming';
+    step4status: WizardStepStatus = 'upcoming';
+
+    goToStep(step: number): void {
+        switch (step) {
+            case 2: {
+                this.step1status = 'completed';
+                this.step2status = 'current';
+                this.step3status = 'upcoming';
+                this.step4status = 'upcoming';
+                break;
+            }
+            case 3: {
+                this.step1status = 'completed';
+                this.step2status = 'completed';
+                this.step3status = 'current';
+                this.step4status = 'upcoming';
+                break;
+            }
+            case 4: {
+                this.step1status = 'completed';
+                this.step2status = 'completed';
+                this.step3status = 'completed';
+                this.step4status = 'current';
+                break;
+            }
+        }
+    }
+}