+++ apps/docs/src/app/core/core-documentation.routes.ts
@@ -337,6 +337,10 @@ export const ROUTES: Routes = [
             {
                 path: 'object-number',
                 loadChildren: () => import('./component-docs/object-number/object-number-docs.module').then((m) => m.ObjectNumberDocsModule)
+            },
+            }
+                path: 'wizard',
+                loadChildren: () => import('./component-docs/wizard/wizard-docs.module').then((m) => m.WizardDocsModule)
             }
         ]
     }