+++ apps/docs/src/app/core/documentation/core-documentation.component.ts
@@ -68,8 +68,8 @@ export class CoreDocumentationComponent extends DocumentationBaseComponent {
             { url: 'core/switch', name: 'Switch' },
             { url: 'core/token', name: 'Token' },
             { url: 'core/toolbar', name: 'Toolbar' },
-            { url: 'core/object-number', name: 'Object Number' }
-            // { url: 'core/tree', name: 'Tree' }
+            { url: 'core/object-number', name: 'Object Number' },
+            { url: 'core/wizard', name: 'Wizard' }
         ];
 
         this.layouts = [