+++ libs/core/src/index.ts
@@ -57,6 +57,7 @@ export * from './lib/toolbar/public_api';
 export * from './lib/tree/public_api';
 export * from './lib/info-label/public_api';
 export * from './lib/object-status/public_api';
+export * from './lib/wizard/public_api';
 export * from './lib/object-identifier/public_api';
 export * from './lib/card/public_api';
 export * from './lib/object-number/public_api';